+++ example/lib/example/example.dart
@@ -1,0 +1,3 @@
+export 'example_1.dart';
+export 'example_2.dart';
+export 'example_3.dart';